+++ tools/src/librlink/ReventFd.hpp
@@ -1,6 +1,6 @@
-// $Id: ReventFd.hpp 983 2018-01-02 20:35:59Z mueller $
+// $Id: ReventFd.hpp 1084 2018-12-16 12:23:53Z mueller $
 //
-// Copyright 2013- by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+// Copyright 2013-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
 // This program is free software; you may redistribute and/or modify it under
 // the terms of the GNU General Public License as published by the Free
@@ -13,6 +13,7 @@
 // 
 // Revision History: 
 // Date         Rev Version  Comment
+// 2018-12-16  1084   1.0.1  use =delete for noncopyable instead of boost
 // 2013-01-14   475   1.0    Initial version
 // 2013-01-11   473   0.5    First draft
 // ---------------------------------------------------------------------------
@@ -26,14 +27,15 @@
 #ifndef included_Retro_ReventFd
 #define included_Retro_ReventFd 1
 
-#include "boost/utility.hpp"
-
 namespace Retro {
 
-  class ReventFd : private boost::noncopyable {
+  class ReventFd {
     public:
                     ReventFd();
       virtual      ~ReventFd();
+
+                    ReventFd(const ReventFd&) = delete;   // noncopyable 
+      ReventFd&     operator=(const ReventFd&) = delete;  // noncopyable
 
       int           Fd() const;
       int           Signal();