+++ tools/src/librtcltools/RtclSetList.hpp
@@ -1,4 +1,4 @@
-// $Id: RtclSetList.hpp 1083 2018-12-15 19:19:16Z mueller $
+// $Id: RtclSetList.hpp 1084 2018-12-16 12:23:53Z mueller $
 //
 // Copyright 2013-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
@@ -13,6 +13,7 @@
 // 
 // Revision History: 
 // Date         Rev Version  Comment
+// 2018-12-16  1084   1.2.3  use =delete for noncopyable instead of boost
 // 2018-12-15  1083   1.2.2  Add(): use rval ref and move semantics
 // 2018-12-14  1081   1.2.1  use std::function instead of boost
 // 2018-12-01  1076   1.2    use unique_ptr
@@ -35,19 +36,20 @@
 #include <map>
 #include <functional>
 
-#include "boost/utility.hpp"
-
 #include "RtclSet.hpp"
 #include "librtcltools/RtclArgs.hpp"
 
 namespace Retro {
 
-  class RtclSetList : private boost::noncopyable {
+  class RtclSetList {
     public:
       typedef std::unique_ptr<RtclSetBase> set_uptr_t;
 
                     RtclSetList();
       virtual      ~RtclSetList();
+
+                    RtclSetList(const RtclSetList&) = delete;   // noncopyable 
+      RtclSetList&  operator=(const RtclSetList&) = delete;     // noncopyable
 
       void          Add(const std::string& name, set_uptr_t&& upset);