+++ tools/src/librw11/Rw11Cntl.hpp
@@ -1,6 +1,6 @@
-// $Id: Rw11Cntl.hpp 983 2018-01-02 20:35:59Z mueller $
+// $Id: Rw11Cntl.hpp 1084 2018-12-16 12:23:53Z mueller $
 //
-// Copyright 2013-2017 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+// Copyright 2013-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
 // This program is free software; you may redistribute and/or modify it under
 // the terms of the GNU General Public License as published by the Free
@@ -13,6 +13,7 @@
 // 
 // Revision History: 
 // Date         Rev Version  Comment
+// 2018-12-16  1084   1.2.3  use =delete for noncopyable instead of boost
 // 2017-04-15   874   1.2.2  NUnit() now pure; add UnitBase()
 // 2017-04-02   865   1.2.1  Dump(): add detail arg
 // 2017-02-04   848   1.2    add ProbeFound(),ProbeDataInt,Rem()
@@ -33,8 +34,6 @@
 
 #include <string>
 
-#include "boost/utility.hpp"
-
 #include "librtools/Rstats.hpp"
 #include "librlink/RlinkConnect.hpp"
 #include "librlink/RlinkServer.hpp"
@@ -47,12 +46,15 @@ namespace Retro {
 
   class Rw11Unit;                           // forw decl to avoid circular incl
 
-  class Rw11Cntl : public Rbits, private boost::noncopyable {
+  class Rw11Cntl : public Rbits {
     public:
 
       explicit      Rw11Cntl(const std::string& type);
       virtual      ~Rw11Cntl();
 
+                    Rw11Cntl(const Rw11Cntl&) = delete;   // noncopyable 
+      Rw11Cntl&     operator=(const Rw11Cntl&) = delete;  // noncopyable
+    
       void          SetCpu(Rw11Cpu* pcpu);
       Rw11Cpu&      Cpu() const;
       Rw11&         W11() const;